pyproject.toml
[tool.poetry]
name = "image-recreation"
version = "0.1.0"
description = "Genetic Algorithm for Image Recreation"
authors = ["Natchy"]
readme = "README.md"

[tool.poetry.dependencies]
python = "<3.14,>=3.10"
uvicorn = "^0.32.1"
fastapi = "^0.115.6"
streamlit = "^1.40.2"
numpy = "^2.2.0"
pillow = "^11.0.0"
matplotlib = "^3.9.3"
ipython = "^8.30.0"
colour = "^0.1.5"
seaborn = "^0.13.2"
colour-science = "^0.4.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
